requirements.txt
python-telegram-bot>=20.8
requests==2.31.0
python-dotenv==1.0.0
pydantic==1.10.8
urllib3==2.0.7 